requirements.txt
pyautoit
pyautogui
keyboard
pywin32
mouse